14. Vertical animations
Vertical animations are added for all the pages. Also, in the app file everything except the nav component is wrapped inside the AnimatePresence component of framer motion, so the app will wait for a transition to finish before statrting another one, i.e. when changing from one page to another the vertical fading effect of one page has to end before the animation of the selected page starts.