17. OurWork page animations
Some styled divs that occupy the whole width and height with different colors, called Frame#, ared added (for an animation). Three new animations are added:

-lineAnim: the line that divides each video's title and photo goes from width 0% to 100%.
-slider: moves x position of frames form left to right, while rotating them from 45 degrees to 0 degrees.
-sliderContainer: Used for staggering the the slider animation of the Frame# elements (colored divs).

A Hide element (styled div with with overflow: hidden) is added to hide the parts of the image that go out of the original image's frame size, because of the photoAnim scaling (looks better).