18. Toggle component for questions in faq page
A reusable component called Toggle is created. This wraps the divs with
the class "question". h4s with the text of the questions are deleted as
the title's text is passed as a prop to the Toggle component, also, the
lines (divs with class "faq-line") are deleted, since those are now
included inside the Toggle component.

The Toggle component uses the children prop to render all the html
placed among the the Toggle component tags - <Toggle>Some html tags
(this is accesed through the children prop)</Toggle> - when the toggle
state is set to true, otherwise just the question title is shown (passed
as the title prop). The toggle state changes from false to true and
viceversa when the the question is clicked.