19. FAQ section animation
The question question class is moved to the outer div inside the Toggle
component. The AnimatedSharedLayout is used for wrapping all the Toggles
so the answers of the question have an animation when shown or hidden;
the layout property is added to the motion.div and motion.h4 inside the
Toggle component.